Add Microsoft's ".NET for Java developers" eBook
Add a link to https://dotnet.microsoft.com/en-us/campaigns/dotnet-for-java-developers, Microsoft's free ".NET for Java developers" eBook.

This eBook is intended to offer a quick introduction to C# and .NET Core for those experienced in the Java ecosystem. The link I'm adding is to their landing page for the book and Microsoft has a download link on that landing page, thankfully one that does not ask for contact information before offering the free eBook.